restful_python/requirements.txt
flask
psycopg2-binary 
gunicorn
toml
